requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns
aws_cdk.aws_logs
aws_cdk.aws_elasticloadbalancingv2
